requirements.txt
asgiref==3.6.0
bleach==5.0.1
click==8.1.3
colorama==0.4.6
Django==4.1.4
django-post-office==3.6.3
djangorestframework==3.14.0
jsonfield==3.1.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.3
pip==22.0.4
platformdirs==2.6.2
python-dotenv==0.21.0
pytz==2022.7
setuptools==58.1.0
six==1.16.0
sqlparse==0.4.3
tinycss2==1.1.1
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
webencodings==0.5.1
